chore: Added tailwind output to design tokens generation.
Added an additional output to the design-tokens output. It sources what we currently have in the
alias definitions and patches it into a tailwindcss configuration, which is then passed to the
tailwind generator.
It ouputs a css file that will be shipped with the design-tokens package and can be used standalone
without tailwindcss or postcss.

Fixes #1182