fix(filter-field): Fixes an issue with the infix when deleting a filter while in flight.
When deleting an already set filter with the mouse, while the filter field is in the middle of
adding a new filter, the prefix was deleted.
There was a reset of the prefix, when a filter was deleted that is not the current filter, which is
removed in this commit. I could not find where this would have caused trouble.
I added a test for the current behaviour and have all other tests passing.

Fixes #1264